python-server/requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
bleach==3.1.0
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.7.1
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
enum34==1.1.6
Flask==1.0.2
functools32==3.2.3.post2
futures==3.2.0
google-api-core==1.7.0
google-auth==1.6.3
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
idna==2.8
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.0
opencv-python==4.0.0.21
pandocfilters==1.4.2
pathlib2==2.3.3
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==1.0.15
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
pyparsing==2.3.1
pytesseract==0.2.6
python-dateutil==2.7.5
pytz==2018.9
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
rsa==4.0
scandir==1.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-unicode-console==0.5
